+++ server/package.json
@@ -5,10 +5,11 @@
   "main": "server.ts",
   "author": "",
   "license": "AGPL-3.0",
+  "type": "module",
   "scripts": {
     "build": "tsc",
     "prestart": "yarn build",
-    "start": "nodemon server.ts",
+    "start": "nodemon src/server.ts",
     "heroku-start": "node dist/server.js",
     "test": "jest -i"
   },
@@ -29,27 +30,30 @@
     "common": "0.0.1",
     "cors": "^2.8.5",
     "dotenv": "^16.4.4",
-    "express": "^4.18.2",
     "google-auth-library": "^6.0.0",
-    "heroku-ssl-redirect": "^0.1.1",
+    "heroku-ssl-redirect": "^0.0.4",
     "joi": "^17.11.0",
     "mongodb": "^3.5.5",
     "mongodb-memory-server": "^6.5.2",
     "mongoose": "^5.9.6",
-    "nodemon": "^3.0.2",
-    "shortid": "^2.2.15",
-    "ts-node": "^10.9.2"
+    "shortid": "^2.2.15"
   },
   "devDependencies": {
     "@types/cors": "^2.8.6",
+    "@types/express": "^4.17.21",
     "@types/jest": "^29.x",
     "@types/mongodb": "^3.5.4",
     "@types/mongoose": "^5.7.7",
+    "@types/node": "^20.11.19",
     "@types/prettier": "<2.6.0",
     "@types/shortid": "^0.0.29",
     "@types/ws": "8.5.4",
+    "express": "^4.18.2",
     "jest": "^29.x",
     "jest-circus": "^29.x",
-    "ts-jest": "^29.x"
+    "nodemon": "^3.0.3",
+    "ts-jest": "^29.x",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
   }
 }
