+++ server/src/admin/index.ts
@@ -1,1 +1,1 @@
-export { adminRouter } from "./admin.router";
+export { adminRouter } from "./admin.router.js";
