+++ server/src/auth/auth.data-access.ts
@@ -1,4 +1,4 @@
-import { Students } from '../../db/schema';
+import { Students } from '../db/schema.js';
 
 export const insertNewStudent = async (studentDoc) => {
   const newStudent = new Students(studentDoc);
