+++ server/src/auth/index.ts
@@ -1,1 +1,1 @@
-export { authRouter } from "./auth.router";
+export { authRouter } from "./auth.router.js";
