+++ server/src/db/index.ts
@@ -9,13 +9,13 @@ const uri = process.env.MONGODB_URL
   : 'Please include a valid MONGODB_URL in the .env file';
 
 export async function setupDb() {
+  console.log(uri)
   try {
     await mongoose.connect(uri, {
       useNewUrlParser: true,
       useUnifiedTopology: true,
     });
   } catch (err) {
-    // eslint-disable-next-line no-console
     console.error('No DB connection defined!');
     process.exit(1);
   }
