+++ server/src/search/index.ts
@@ -1,1 +1,1 @@
-export { searchRouter } from "./search.router";
+export { searchRouter } from "./search.router.js";
