+++ server/src/server.ts
@@ -4,16 +4,20 @@ import path from 'path';
 import express from 'express';
 import sslRedirect from 'heroku-ssl-redirect';
 import cors from 'cors';
+import { fileURLToPath } from "url"
 
-import { setupDb } from './db';
-import { configure } from './endpoints';
+import { setupDb } from './db/index.js';
+import { configure } from './endpoints.js';
 
 const app = express();
+
 app.use(sslRedirect(['development', 'production']));
 
 // technically not needed since we proxy calls?
 app.use(cors());
 
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
 app.use(express.static(path.join(__dirname, '../../client/build')));
 
 function setup() {
