+++ server/src/utils/index.ts
@@ -1,8 +1,8 @@
-export { getCourseById } from '../course/course.controller';
-export { findStudent } from '../profile/profile.data-access';
+export { getCourseById } from '../course/course.controller.js';
+export { findStudent } from '../profile/profile.data-access.js';
 export {
   findReview,
   findReviewCrossListOR,
-} from '../review/review.data-access';
-export { getReviewsCrossListOR } from '../course/course.controller';
-export { insertUser } from '../auth/auth.controller';
+} from '../review/review.data-access.js';
+export { getReviewsCrossListOR } from '../course/course.controller.js';
+export { insertUser } from '../auth/auth.controller.js';
